config.c I_SetField strcat overflowing allocated array space
The code attempts to append Name and Section name, adding a ":" in between.
The following line:
  FullName[Len] = ':';
Will overwrite the NULL terminator byte inserted by the first strcat,
causing the second strcat to start inserting at the wrong position.
This is undefined behaviour, the particular manifestation addressed is
memory corruption error detected by glibc in a later call to malloc.

Signed-off-by: Casey Dwyer <caseydwyer@gmail.com>